refactor and drop online wind data acquisition
