+++ .gitignore
@@ -7,8 +7,4 @@ trial.py
 Wind/backup/2024Dec.nc
 Wind/backup/2024Nov.nc
 Land/raw/
-Wind/backup/2024Sep.nc
-Wind/backup/2024Oct.nc
-Wind/backup/2024Jun.nc
-Wind/backup/2024Jul.nc
-Wind/backup/2024Aug.nc
+Wind/raw/
