+++ Optimiser/config.py
@@ -4,7 +4,6 @@ This is the configuration file and contains most of the hyperparameters.
 
 import os
 import numpy as np
-from Wind.main import wind
 
 '''
 wind farm data
