+++ streamlit_app.py
@@ -127,7 +127,7 @@ if submit:
         feasible_cell = st.session_state['history']['feasible_cell']
     else:
         conv = CRSConvertor([site[1][0], site[0][1], site[0][0], site[1][1]])
-        wind_data = wind([site[1][0], site[0][1], site[0][0], site[1][1]], 'Wind/backup/summary-1d.nc', ['2024'], ['12'], True, True)
+        wind_data = wind([site[1][0], site[0][1], site[0][0], site[1][1]], 'Wind/data/summary-1d.nc')
         feasible_cell = land('Land/data/infeasible.nc', conv.grid_gcs)
     solution, summary, efficiency, st.session_state['wt_summary'] = optimisation(st.session_state['wt_number'], conv.rows, conv.cols, wind_data, feasible_cell)
     solution = conv.gene_to_pos(solution)
